@@ -1,6 +1,6 @@
 {
     "name": "fetch-blob",
-    "version": "3.1.3",
+    "version": "3.1.4",
     "description": "Blob & File implementation in Node.js, originally from node-fetch.",
     "main": "index.js",
     "type": "module",
@@ -14,10 +14,9 @@
         "streams.cjs"
     ],
     "scripts": {
-        "test-wpt": "node --experimental-loader ./test/http-loader.js ./test/test-wpt-in-node.js",
-        "test": "ava test.js",
-        "report": "c8 --reporter json --reporter text ava test.js",
-        "coverage": "c8 --reporter json --reporter text ava test.js && codecov -f coverage/coverage-final.json",
+        "test": "node --experimental-loader ./test/http-loader.js ./test/test-wpt-in-node.js",
+        "report": "c8 --reporter json --reporter text npm run test",
+        "coverage": "npm run report && codecov -f coverage/coverage-final.json",
         "prepublishOnly": "tsc --declaration --emitDeclarationOnly --allowJs index.js from.js"
     },
     "repository": "https://github.com/node-fetch/fetch-blob.git",
@@ -36,11 +35,9 @@
     },
     "homepage": "https://github.com/node-fetch/fetch-blob#readme",
     "devDependencies": {
-        "ava": "^3.15.0",
-        "c8": "^7.7.2",
-        "codecov": "^3.8.2",
-        "node-fetch": "^3.0.0-beta.9",
-        "typescript": "^4.3.2"
+        "@types/node": "^17.0.9",
+        "c8": "^7.11.0",
+        "typescript": "^4.5.4"
     },
     "funding": [
         {
@@ -53,6 +50,7 @@
         }
     ],
     "dependencies": {
+        "node-domexception": "^1.0.0",
         "web-streams-polyfill": "^3.0.3"
     }
 }
