@@ -9,12 +9,19 @@ import * as fg from "./owinteractions/funcGestures.cjs";
 import * as zipgest from "./utils/zipGestures.cjs";
 //const logger = require("./utils/logger.cjs");
 import * as logger from "./utils/logger.cjs";
-
-//const express = require("express");
 import express from 'express';
+import fetch from 'node-fetch';
+//const conf = require("../config/conf.cjs");
+import * as conf from '../config/conf.cjs';
+
+import * as https from 'https';
+
 const app = express();
 app.use(express.json());
 export default app;
+
+const __dirname = path.resolve();
+
 
 
 app.get("/",(req,res)=>{
@@ -40,7 +47,7 @@ app.post("/api/v1/action/merge", async (req, res) => {
     })*/
 
     fg.getAction(sequenceName).then((result) => {
-        var promises = [];
+        var promises = [];blocking
         if (result.toString().includes("OpenWhiskError")) {
             res.json({ mex: result });
             return;
@@ -71,9 +78,6 @@ app.post("/api/v1/action/merge", async (req, res) => {
             /*
                 BISOGNA INVOCARE LA SEQUENZA APPENA FATTA E PRENDERNE LE METRICHE PER VEDERE SE IL MERGE È STATO EFFICACE
                 SE SI POSSO CREARE LA NUOVA FUNZIONE
-
-
-                POSSO EVITARE LA DELETE E FARE LA CREATE CON OVERWRITE == TRUE??
             */
             fg.deleteAction(sequenceName).then(()=>{
                 fg.createAction(sequenceName,wrappedFunc).then(()=>{
@@ -88,7 +92,7 @@ app.post("/api/v1/action/merge", async (req, res) => {
         });
     });
 });
-/*
+
 app.post("/api/v1/listFields", (req, res) => {
 
     var keys = utils.getBodyFields(req.body);
@@ -97,12 +101,30 @@ app.post("/api/v1/listFields", (req, res) => {
     });
 
     res.json({ mex: "END" });
-});*/
+});
 
 app.get("/api/v1/action/list", (req, res) => {
 
-    fg.listAction().then((result) => {
-        res.json({ mex: result });
+    
+    const httpsAgent = new https.Agent({
+        rejectUnauthorized: false,
+      });
+    
+      
+    console.log(__dirname);
+    fetch('https://'+conf.API_HOST+'/api/v1/namespaces/_/actions',{
+        headers: {
+            'Authorization':'Basic '+ btoa(conf.API_KEY)
+        },
+        connection: {
+            ca:fs.readFileSync(path.join(__dirname+'/key.pem'))
+        },
+        agent: httpsAgent
+    })
+    .then(response => response.json())
+    .then(data => {
+        res.json(data);
+        logger.log("/api/v1/action/list" + data,"info")
     });
 
 });
@@ -111,53 +133,98 @@ app.post("/api/v1/action/invoke", (req, res) => {
 
     const funcName = req.body.name;
     logger.log("/api/v1/action/invoke","info");
-    fg.invokeAction(funcName).then((result) => {
-        res.json({ mex: result });
-    });
+    fetch('https://'+conf.API_HOST+'/api/v1/namespaces/_/actions/'+funcName+'?blocking=true',{
+        headers: {
+            'Authorization':'Basic '+ btoa(conf.API_KEY)
+        }
+    })
+    .then(response => response.json())
+    .then(data => logger.log("/api/v1/action/list" + data,"info"));
 
 });
 
 app.post("/api/v1/action/invoke-with-params", (req, res) => {
 
-    const funcName = req.body.name;
+    
     logger.log("/api/v1/action/invoke","info");
 
-    fg.invokeActionWithParams(funcName,req.body.params).then((result) => {
-        res.json({ mex: result });
-    });
+    (async () => {
+        const rawResponse = await fetch('https://'+conf.API_HOST+'/api/v1/namespaces/_/actions/'+req.body.name+'?blocking=true', {
+          method: 'POST',
+          headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json',
+            'Authorization':'Basic '+ btoa(conf.API_KEY)
+          },
+          body: JSON.stringify(req.body.params)
+        });
+        const content = await rawResponse.json();
+      
+        console.log(content);
+      })()
 
 });
 
 app.post("/api/v1/action/get", (req, res) => {
 
+ 
+    const httpsAgent = new https.Agent({
+        rejectUnauthorized: false,
+      });
+
     logger.log("/api/v1/action/get","info");
-    const funcName = req.body.name;
-    fg.getAction(funcName).then((result) => {
-        res.json({ mex: result });
+    fetch('https://'+conf.API_HOST+'/api/v1/namespaces/_/actions/'+req.body.name+'?blocking=true',{
+        method: 'GET',
+        headers: {
+            'Authorization':'Basic '+ btoa(conf.API_KEY)
+        },
+        agent: httpsAgent
+      })
+    .then(response => response.json())
+    .then(data => {
+        res.json(data);
+        logger.log("/api/v1/action/list" + data,"info")
     });
 
 });
 
 app.post("/api/v1/action/delete", (req, res) => {
 
+    // come impostarla con API rest???
     const funcName = req.body.name;
     logger.log("/api/v1/action/delete","info");
-    fg.deleteAction(funcName).then((result) => {
-        res.json({ mex: result });
-    });
+    fetch('https://'+conf.API_HOST+'/api/v1/namespaces/_/actions/'+req.body.name,{
+        method: 'DELETE',
+        headers: {
+            'Authorization':'Basic '+ btoa(conf.API_KEY)
+        }
+      })
+    .then(response => response.json())
+    .then(data => logger.log("/api/v1/action/list" + data,"info"));
+
+
 
 });
 
 app.post("/api/v1/action/create", (req, res) => {
 
     logger.log("/api/v1/action/create","info");
-    const funcName = req.body.name;
-    const funcBody = req.body.fbody;
     
-    fg.createAction(funcName, funcBody).then((result) => {
-        res.json({ mex: result });
-    });
-
+    (async () => {
+        const rawResponse = await fetch('https://'+conf.API_HOST+'/api/v1/namespaces/_/actions/'+req.body.name+'?overwrite=true', {
+          method: 'PUT',
+          headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json',
+            'Authorization':'Basic '+ btoa(conf.API_KEY)
+          },
+          // posso evitare di specirficare "kind"?
+          body: JSON.stringify({"namespace":"_","name":req.body.name,"exec":{"kind":"nodejs:10","code":req.body.fbody},"annotations":[{"key":"web-export","value":true},{"key":"raw-http","value":false},{"key":"final","value":true}]})
+        });
+        const content = await rawResponse.json();
+      
+        console.log(content);
+      })()
 });
 
 function parseFunction(element,timestamp){
